+++ threedi_schema/migrations/versions/0223_upgrade_db_inflow.py
@@ -78,7 +78,7 @@ ADD_TABLES = {
         Column("tags", Text),
         Column("distribution", Text)
     ],
-    "tags": [
+    "tag": [
         Column("description", Text)
     ]
 }