+++ threedi_schema/__init__.py
@@ -2,6 +2,6 @@ from .application import *  # NOQA
 from .domain import constants, custom_types, models  # NOQA
 
 # fmt: off
-__version__ = '0.228.1.dev0'
+__version__ = '0.228.1'
 
 # fmt: on