+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog of threedi-schema
 0.228.1 (unreleased)
 --------------------
 
+- Rename sqlite table "tags" to "tag"
 - Remove indices referring to removed tables in previous migrations
 - Remove columns referencing v2 in geometry_column